README.md
Hello! This You Are An Idiot popup was made by me. I wanted to prank my friends with this, a file called "hoi4.exe". It's been a while, so I thought I'd put it up. I made it in Python. Once you've started it, press F12 and type "IMIDIOT" to exit. Have fun! :)